+++ packages/article-magazine-comment/__tests__/web/semantic.web.test.js
@@ -117,10 +117,8 @@ const tests = [
               brightcovePlayerId: "default",
               caption: "This is video caption",
               display: "primary",
-              paidOnly: "false",
               posterImageId: "0c0309d4-1aeb-11e8-9010-1eef6ba5d3de",
-              posterImageUrl: "https://image.io",
-              skySports: false
+              posterImageUrl: "https://image.io"
             },
             children: [],
             name: "video"