+++ packages/article-skeleton/src/article-body/article-body.web.js
@@ -214,8 +214,7 @@ const renderers = ({ paidContentClassName, template }) => ({
       brightcoveAccountId,
       brightcovePlayerId,
       caption,
-      posterImageUrl,
-      skysports
+      posterImageUrl
     }
   ) {
     const MediaWrapper = responsiveDisplayWrapper("primary");
@@ -231,7 +230,6 @@ const renderers = ({ paidContentClassName, template }) => ({
               playerId={brightcovePlayerId}
               policyKey={brightcovePolicyKey}
               poster={{ uri: posterImageUrl }}
-              skySports={skysports}
               videoId={brightcoveVideoId}
               width="100%"
             />