+++ packages/video/__tests__/default-video-props.js
@@ -4,14 +4,12 @@ const videoProps = {
   id: "[tpa video id]",
   is360: false,
   onVideoPress: () => {},
-  paidOnly: false,
   playerId: "default",
   policyKey: "[policy key]",
   poster: {
     uri:
       "data:image/gif;base64,R0lGODlhAQABAIAAAAAAAP///yH5BAEAAAAALAAAAAABAAEAAAIBRAA7"
   },
-  skySports: false,
   videoId: "[video id]",
   width: 300
 };