+++ packages/video/__tests__/video-shared-with-style.native.js
@@ -13,9 +13,6 @@ import Video from "../src/video";
 import defaultVideoProps from "./default-video-props";
 
 jest.mock("@times-components/image", () => "Image");
-jest.mock("@times-components/gradient", () => ({
-  OverlayGradient: "OverlayGradient"
-}));
 
 // eslint-disable-next-line global-require
 jest.mock("@times-components/svgs", () => require("./mock-svg"));
@@ -46,16 +43,6 @@ export default () => {
       test: () => {
         const testInstance = TestRenderer.create(
           <Video {...defaultVideoProps} poster={null} />
-        );
-
-        expect(testInstance).toMatchSnapshot();
-      }
-    },
-    {
-      name: "sky sports video",
-      test: () => {
-        const testInstance = TestRenderer.create(
-          <Video {...defaultVideoProps} skySports />
         );
 
         expect(testInstance).toMatchSnapshot();