+++ packages/video/__tests__/web/video-with-data-attributes.web.test.js
@@ -41,7 +41,7 @@ const tests = [
     test: () => {
       const testInstance = TestRenderer.create(
         <IsPaidSubscriber.Provider value>
-          <Video {...defaultVideoProps} paidOnly={false} poster={null} />
+          <Video {...defaultVideoProps} poster={null} />
         </IsPaidSubscriber.Provider>
       );
 
@@ -53,7 +53,7 @@ const tests = [
     test: () => {
       const testInstance = TestRenderer.create(
         <IsPaidSubscriber.Provider value>
-          <Video {...defaultVideoProps} paidOnly={false} poster={null} is360 />
+          <Video {...defaultVideoProps} poster={null} is360 />
         </IsPaidSubscriber.Provider>
       );