+++ packages/video/src/video-prop-types.js
@@ -10,19 +10,15 @@ export const propTypes = {
   height: numberOrString.isRequired,
   id: PropTypes.string,
   is360: PropTypes.bool,
-  paidOnly: PropTypes.oneOfType([PropTypes.string, PropTypes.bool]),
   playerId: PropTypes.string,
   policyKey: PropTypes.string.isRequired,
   poster: PropTypes.shape({ uri: PropTypes.string.isRequired }),
-  skySports: PropTypes.bool,
   videoId: PropTypes.string.isRequired,
   width: numberOrString.isRequired
 };
 
 export const defaultProps = {
   is360: false,
-  paidOnly: false,
   playerId: "default",
-  poster: null,
-  skySports: false
+  poster: null
 };