+++ packages/article-main-standard/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.26.15](https://github.com/newsuk/times-components/compare/@times-components/article-main-standard@3.26.14...@times-components/article-main-standard@3.26.15) (2020-03-12)
+
+
+### Bug Fixes
+
+* Remove paidOnly and skySports video attributes ([#2555](https://github.com/newsuk/times-components/issues/2555)) ([aff064c](https://github.com/newsuk/times-components/commit/aff064c06c25062763669fe29c6afe124d57803c))
+
+
+
+
+
 ## [3.26.14](https://github.com/newsuk/times-components/compare/@times-components/article-main-standard@3.26.13...@times-components/article-main-standard@3.26.14) (2020-03-10)
 
 **Note:** Version bump only for package @times-components/article-main-standard