+++ packages/article-skeleton/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.32.4](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.32.3...@times-components/article-skeleton@1.32.4) (2020-03-12)
+
+
+### Bug Fixes
+
+* Remove paidOnly and skySports video attributes ([#2555](https://github.com/newsuk/times-components/issues/2555)) ([aff064c](https://github.com/newsuk/times-components/commit/aff064c06c25062763669fe29c6afe124d57803c))
+
+
+
+
+
 ## [1.32.3](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.32.2...@times-components/article-skeleton@1.32.3) (2020-03-10)
 
 **Note:** Version bump only for package @times-components/article-skeleton