+++ packages/icons/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.14.38](https://github.com/newsuk/times-components/compare/@times-components/icons@2.14.37...@times-components/icons@2.14.38) (2020-03-12)
+
+**Note:** Version bump only for package @times-components/icons
+
+
+
+
+
 ## [2.14.37](https://github.com/newsuk/times-components/compare/@times-components/icons@2.14.36...@times-components/icons@2.14.37) (2020-03-10)