+++ packages/pagination/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.4.22](https://github.com/newsuk/times-components/compare/@times-components/pagination@3.4.21...@times-components/pagination@3.4.22) (2020-03-12)
+
+**Note:** Version bump only for package @times-components/pagination
+
+
+
+
+
 ## [3.4.21](https://github.com/newsuk/times-components/compare/@times-components/pagination@3.4.20...@times-components/pagination@3.4.21) (2020-03-10)
 
 **Note:** Version bump only for package @times-components/pagination