+++ packages/video-label/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.39](https://github.com/newsuk/times-components/compare/@times-components/video-label@2.4.38...@times-components/video-label@2.4.39) (2020-03-12)
+
+**Note:** Version bump only for package @times-components/video-label
+
+
+
+
+
 ## [2.4.38](https://github.com/newsuk/times-components/compare/@times-components/video-label@2.4.37...@times-components/video-label@2.4.38) (2020-03-10)
 
 **Note:** Version bump only for package @times-components/video-label